+++ CHANGELOG.md
@@ -78,3 +78,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [1.1.0] - 2025-12-07
 
 - Included new Output variable ecs_container_name
+
+## [1.1.1] - 2025-12-07
+- Added Output variables for security group ids 
+- Introduced allowed CIDR blocks variable